FEATURE: support consistent policy renewal dates
Previous to this change, policy renewal was solely conditional on
policy creation date. This makes it very awkward to have policies that
remind people of stuff every quarter cause the date is constantly floating

With this change you can use `renew-start="01-01-2010"` to determine
the first date a renewal happens, after which repeat renewals will only
be queue if N*renew-days + renew-start has passed.

This also is a major refactor of all policy internals, we no longer
use a mountain of post custom fields, instead we have structured
data in the post_policies table.

Yes, the migration is hairy.

Future plans include also moving the policy acceptance into a dedicated
table so we can keep track of policy acceptance history.