UX: When a policy reminder notification is sent mark it as high priority (#12)
This is so policy reminders do not fall by the wayside. They are given the same priority as bookmark reminder and PM notifications, so they show with the green notification bubble.

Any existing reminder notification will be deleted as well, we don't want the user to end up with tons of duplicate notifications if the policy reminder is aggressive.