FEATURE: Track when a policy version is bumped
Currently we have queries to check who hasn’t accepted a policy and when
that policy was last triggered for that person.
The problem is that it’s not always reliable as we don’t have a way of
knowing when the policy version was bumped for the last time.

This patch adds a new `last_bumped_at` field which will be updated
automatically each time the version changes. That way we’ll be able to
have more reliable queries.